move exponential_search to binary_search.py to pass github auto build tests
delete exponential_search.py file
